Helper

Description

Documentation

fst' :: (a, b, c) -> a

returns the first element of a 3-element-tupel

snd' :: (a, b, c) -> b

returns the second element of a 3-element-tupel

thd' :: (a, b, c) -> c

returns the third element of a 3-element-tupel

prompt :: String -> IO String

helper function for reading command line.

validPort :: String -> Bool

checks if the entered port is a valid port

onlyClient :: String -> Bool

checks if the user wants to create or just join a game.

intArrayToString :: [Int] -> String

formats a list of integer to a string