Message

Description

Documentation

messageHandler :: Text -> String

converts command line input into json.

jsonStringify :: ByteString -> String

takes a json object and convert it to a string.

jsonParse :: String -> ByteString

takes a string and converts it to a json object.

jsonToMessageContainer :: ByteString -> Maybe Message

takes a jsonObject and puts it in a container.

getCommandOfMessage :: Message -> String

returns the parameter "command" of a message.

getParameterOfMessage :: Message -> String

returns the parameter "parameter" of a message.

extractContainer :: Maybe a -> a

takes Maybe a and returns a.

data Message

Message data.