MianetDistributor

Description

Documentation

sendToSenderClient :: Connection -> Text -> IO ()

sends a message to the last active client.

sendToActor :: MVar ServerState -> Text -> IO ()

sends a message to the actor client.

sendToReactor :: MVar ServerState -> Text -> IO ()

sends a message to the reactor client.

sendToAllClients :: MVar ServerState -> Text -> IO ()

sends a message to all clients

sendToAllClientsExceptSender :: Client -> MVar ServerState -> Text -> IO ()

sends a message to all clients except of the sender client (last active client).

sendToAllExceptActor :: MVar ServerState -> Text -> IO ()

sends a message to all clients except of the actor client.

sendToAllExceptActorAndReactor :: MVar ServerState -> Text -> IO ()

sends a message to all clients except of actor and reactor.

sendToWinner :: MVar ServerState -> Int -> Text -> IO ()

sends a message to winners (winner client).

sendToLooser :: MVar ServerState -> Int -> Text -> IO ()

send a message to the loosers.