MianetDraws

Description

Documentation

moveClient :: ServerState -> ServerState

moves the first client in the connection list to the last position.

accuse :: MVar ServerState -> MVar Draw -> Int -> IO ()

the reactor accused the actor of lying. it returns the result.

incrementScore :: MVar ServerState -> (MVar ServerState -> IO Text) -> IO ()

increments the score of the actor or reactor.

incrementWinCountForClient :: Text -> ServerState -> ServerState

increments the score of a specific client.

nextDraw :: MVar ServerState -> MVar Draw -> Int -> Int -> IO ()

initiates the next draw and sends the respective messages to the clients.

moveToNextDraw :: MVar ServerState -> IO (MVar ServerState)

initiates the next draw. it calls the moveClient function.

rollDices' :: MVar Draw -> IO Int

rolls dices two times and returns the result.

checkForWin :: MVar ServerState -> Int -> IO Bool

checks if somebody got the max score.

endOfGameFunc :: MVar ServerState -> Int -> IO ()

inititates the end of a game.

logResult :: MVar ServerState -> MVar Draw -> String -> IO ()

it logs the entered dice result of the actor.

validResult :: String -> Bool

checks if it's a valid result.

addToListWithConv :: [Int] -> Char -> [Int]

converts a char to a list of integer and puts it to another list of integer.

addToListWithConv' :: [Int] -> Char -> Maybe [Int]

in process.